Move UI-related sagas
