+++ src/main.js
@@ -2,7 +2,6 @@ import { setupDevelopmentTools } from './main/dev';
 import { setupErrorHandling } from './main/errors';
 import { handleStartup } from './main/startup';
 import { setupAppEvents } from './main/events';
-import { setupRootWindow } from './main/rootWindow';
 import { setupReduxStore } from './main/reduxStore';
 
 if (require.main === module) {
@@ -11,5 +10,4 @@ if (require.main === module) {
 	handleStartup();
 	setupReduxStore();
 	setupAppEvents();
-	setupRootWindow();
 }
