+++ src/main/channels/index.js
@@ -1,6 +1,5 @@
-import { eventChannel, channel } from 'redux-saga';
-
-import { getStore } from '../reduxStore';
+import { eventChannel, channel, buffers } from 'redux-saga';
+import { app } from 'electron';
 
 export const eventEmitterChannel = (eventEmitter, eventName) => eventChannel((emit) => {
 	eventEmitter.addListener(eventName, emit);
@@ -10,30 +9,57 @@ export const eventEmitterChannel = (eventEmitter, eventName) => eventChannel((em
 	};
 });
 
-export const storeChangeChannel = (selector, equalityFunction = Object.is) => {
+const READY = Symbol('READY');
+
+export const appReadyChannel = () => {
 	const chan = channel();
 
-	let closed = false;
-	let unsubscribe = () => undefined;
+	app.whenReady().then(() => {
+		chan.put(READY);
+	});
 
-	getStore().then((store) => {
-		if (closed) {
-			return;
+	return chan;
+};
+
+export const storeValueChannel = (store, selector, equalityFunction = Object.is) => {
+	const chan = channel(buffers.expanding());
+
+	let prevValue;
+	const unsubscribe = store.subscribe(() => {
+		const value = selector(store.getState());
+
+		if (!equalityFunction(value, prevValue)) {
+			chan.put(value);
+			prevValue = value;
 		}
-
-		let prevValue;
-		unsubscribe = store.subscribe(() => {
-			const value = selector(store.getState());
-
-			if (!equalityFunction(value, prevValue)) {
-				chan.put([value, prevValue]);
-				prevValue = value;
-			}
-		});
 	});
 
 	const close = () => {
-		closed = true;
+		unsubscribe();
+		chan.close();
+	};
+
+	return {
+		take: chan.take,
+		flush: chan.flush,
+		close,
+	};
+};
+
+export const storeChangeChannel = (store, selector, equalityFunction = Object.is) => {
+	const chan = channel(buffers.expanding());
+
+	let prevValue;
+	const unsubscribe = store.subscribe(() => {
+		const value = selector(store.getState());
+
+		if (!equalityFunction(value, prevValue)) {
+			chan.put([value, prevValue]);
+			prevValue = value;
+		}
+	});
+
+	const close = () => {
 		unsubscribe();
 		chan.close();
 	};
