+++ src/main/reduxStore.js
@@ -3,19 +3,7 @@ import { createStore, applyMiddleware } from 'redux';
 import createSagaMiddleware from 'redux-saga';
 
 import { rootReducer } from '../reducers';
-
-let resolveStore = () => undefined;
-let resolveSagaMiddleware = () => undefined;
-
-const deferredStore = new Promise((resolve) => {
-	resolveStore = resolve;
-});
-
-export const getStore = () => deferredStore;
-
-const deferredSagaMiddleware = new Promise((resolve) => {
-	resolveSagaMiddleware = resolve;
-});
+import { rootSaga } from './sagas';
 
 export const setupReduxStore = () => {
 	const sagaMiddleware = createSagaMiddleware();
@@ -23,13 +11,9 @@ export const setupReduxStore = () => {
 	const store = createStore(rootReducer, {}, middlewares);
 	replayActionMain(store);
 
-	resolveStore(store);
-	resolveSagaMiddleware(sagaMiddleware);
+	sagaMiddleware.setContext({
+		store,
+		run: sagaMiddleware.run,
+	});
+	sagaMiddleware.run(rootSaga);
 };
-
-export const dispatch = (action) => deferredStore.then((store) => store.dispatch(action));
-
-export const runSaga = (saga, ...args) =>
-	deferredSagaMiddleware.then((sagaMiddleware) =>
-		sagaMiddleware.run(saga, ...args),
-	);
